move sol_compat to elf_loader.rs
