+++ src/lib.rs
@@ -744,39 +744,6 @@ pub unsafe extern "C" fn sol_compat_instr_execute_v1(
     1
 }
 
-#[no_mangle]
-pub unsafe extern "C" fn sol_compat_elf_loader_v1(
-    out_ptr: *mut u8,
-    out_psz: *mut u64,
-    in_ptr: *mut u8,
-    in_sz: u64,
-) -> c_int { 
-    let in_slice = std::slice::from_raw_parts(in_ptr, in_sz as usize);
-    let elf_loader_ctx = match proto::ElfLoaderCtx::decode(in_slice) {
-        Ok(context) => context,
-        Err(_) => return 0,
-    };
-    let elf_bytes = match elf_loader_ctx.elf {
-        Some(elf) => elf.data,
-        None => return 0,
-    };
-
-    let elf_loader_effects = match elf_loader::load_elf(
-        elf_bytes.as_slice(),
-    ) {
-        Some(v) => v,
-        None => return 0,
-    };
-    let out_slice = std::slice::from_raw_parts_mut(out_ptr, (*out_psz) as usize);
-    let out_vec = elf_loader_effects.encode_to_vec();
-    if out_vec.len() > out_slice.len() {
-        return 0;
-    }
-    out_slice[..out_vec.len()].copy_from_slice(&out_vec);
-    *out_psz = out_vec.len() as u64;
-    1
-}
-
 
 #[cfg(test)]
 mod tests {
